+++ src/Components/Home/Home.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './Home.css';
+import '../CSS/Home.css';
 
 
 const Home = () => {
