+++ src/Components/Quotes/Quotes.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import Card from './Card';
-import './Quotes.css';
+import Card from './Card.js';
+import '../CSS/Quotes.css';
 import { Link } from 'react-router-dom';
